# Performance optimization for MultiProducerStrategy