--- a/core/test/CMakeLists.txt
+++ b/core/test/CMakeLists.txt
@@ -32,6 +32,7 @@ function(add_app_test TEST_NAME)
 endfunction()
 
 add_ut_test(qa_buffer)
+add_ut_test(qa_AtomicBitset)
 add_ut_test(qa_DynamicBlock)
 add_ut_test(qa_DynamicPort)
 add_ut_test(qa_HierBlock)