… core/include/gnuradio-4.0/ClaimStrategy.hpp
… core/test/qa_buffer.cpp