Performance optimization for MultiProducerStrategy
* AtomicBitset bulk set/reset
* Rename to Single(Multi)ProducerStrategy
* Eliminate usage of reserve(), use tryReserve() instead
* Remove warnings


### Some results of `bm_Buffer test.

#### AtomicBitset without bulk set/reset
```text
┌──────────────────benchmark:──────────────────┬──────┬──CPU branch misses───┬─ops/s─┬──mean──┬─<CPU-I>─┬───CPU cache misses────┬─stddev─┬─#N─┬─CTX-SW─┬─total time─┬──min───┬─median─┬──max───┐
│                                              │ SKIP │                      │       │        │         │                       │        │    │        │            │        │        │        │
│    POSIX: 1 producers -< 1  >-> 1 consumers  │ PASS │ 63.3k / 546k = 11.6% │ 25.8M │  39 ms │    317m │  13.3k / 100k = 13.2% │   8 ms │ 8  │   105  │     310 ms │  31 ms │  36 ms │  58 ms │
│    POSIX: 1 producers -< 1  >-> 2 consumers  │ PASS │ 80.0k / 606k = 13.2% │ 16.7M │  60 ms │    352m │  71.5k / 170k = 42.0% │   4 ms │ 8  │   112  │     478 ms │  54 ms │  61 ms │  64 ms │
│    POSIX: 1 producers -< 1  >-> 4 consumers  │ PASS │ 84.8k / 622k = 13.6% │ 11.3M │  88 ms │    362m │  87.3k / 194k = 44.9% │   5 ms │ 8  │   120  │     706 ms │  81 ms │  91 ms │  93 ms │
│    POSIX: 2 producers -< 1  >-> 1 consumers  │ PASS │ 92.9k / 678k = 13.7% │  6.8M │ 146 ms │    397m │   104k / 230k = 45.1% │  16 ms │ 8  │   132  │       1  s │ 129 ms │ 142 ms │ 183 ms │
│    POSIX: 2 producers -< 1  >-> 2 consumers  │ PASS │ 95.6k / 707k = 13.5% │  6.4M │ 156 ms │    412m │  94.0k / 218k = 43.1% │  17 ms │ 8  │   140  │       1  s │ 138 ms │ 157 ms │ 184 ms │
│    POSIX: 2 producers -< 1  >-> 4 consumers  │ PASS │  114k / 848k = 13.4% │  6.2M │ 161 ms │    493m │   102k / 248k = 41.2% │  18 ms │ 8  │   164  │       1  s │ 138 ms │ 154 ms │ 196 ms │
│    POSIX: 4 producers -< 1  >-> 1 consumers  │ PASS │  108k / 814k = 13.3% │  5.5M │ 183 ms │    473m │  94.1k / 243k = 38.7% │   5 ms │ 8  │   172  │       1  s │ 177 ms │ 183 ms │ 191 ms │
│    POSIX: 4 producers -< 1  >-> 2 consumers  │ PASS │  104k / 774k = 13.5% │  5.2M │ 191 ms │    450m │   108k / 264k = 40.9% │   3 ms │ 8  │   176  │       2  s │ 184 ms │ 192 ms │ 194 ms │
│    POSIX: 4 producers -< 1  >-> 4 consumers  │ PASS │ 170k / 1.22M = 14.0% │  3.5M │ 282 ms │    711m │   206k / 466k = 44.2% │  16 ms │ 8  │   296  │       2  s │ 256 ms │ 282 ms │ 305 ms │
├──────────────────────────────────────────────┼──────┼──────────────────────┼───────┼────────┼─────────┼───────────────────────┼────────┼────┼────────┼────────────┼────────┼────────┼────────┤
│    POSIX: 1 producers -<1024>-> 1 consumers  │ PASS │ 29.6k / 263k = 11.3% │  8.2G │ 122 us │    153m │ 1.47k / 7.21k = 20.4% │ 154 us │ 8  │     5  │     976 us │  31 us │  32 us │ 506 us │
│    POSIX: 1 producers -<1024>-> 2 consumers  │ PASS │ 33.9k / 303k = 11.2% │  2.3G │ 432 us │    177m │ 3.82k / 13.4k = 28.4% │ 987 us │ 8  │     8  │       3 ms │  40 us │  44 us │   3 ms │
│    POSIX: 1 producers -<1024>-> 4 consumers  │ PASS │ 47.0k / 405k = 11.6% │  3.7G │ 273 us │    237m │ 1.68k / 14.4k = 11.6% │ 213 us │ 8  │    13  │       2 ms │ 156 us │ 158 us │ 815 us │
│    POSIX: 2 producers -<1024>-> 1 consumers  │ PASS │ 60.1k / 525k = 11.4% │  144M │   7 ms │    306m │ 2.72k / 61.2k =  4.4% │  34 us │ 8  │    72  │      56 ms │   7 ms │   7 ms │   7 ms │
│    POSIX: 2 producers -<1024>-> 2 consumers  │ PASS │ 60.1k / 525k = 11.4% │  144M │   7 ms │    306m │ 2.97k / 62.0k =  4.8% │ 101 us │ 8  │    72  │      55 ms │   7 ms │   7 ms │   7 ms │
│    POSIX: 2 producers -<1024>-> 4 consumers  │ PASS │ 60.6k / 529k = 11.5% │  146M │   7 ms │    308m │ 3.25k / 63.6k =  5.1% │ 303 us │ 8  │    72  │      55 ms │   6 ms │   7 ms │   7 ms │
│    POSIX: 4 producers -<1024>-> 1 consumers  │ PASS │ 59.1k / 515k = 11.5% │  217M │   5 ms │    300m │ 2.45k / 56.2k =  4.4% │  15 us │ 8  │    64  │      37 ms │   5 ms │   5 ms │   5 ms │
│    POSIX: 4 producers -<1024>-> 2 consumers  │ PASS │ 58.4k / 509k = 11.5% │  220M │   5 ms │    297m │ 1.84k / 54.5k =  3.4% │  87 us │ 8  │    64  │      36 ms │   4 ms │   5 ms │   5 ms │
│    POSIX: 4 producers -<1024>-> 4 consumers  │ PASS │ 58.4k / 510k = 11.5% │  218M │   5 ms │    297m │ 1.89k / 55.2k =  3.4% │  10 us │ 8  │    64  │      37 ms │   5 ms │   5 ms │   5 ms │
├──────────────────────────────────────────────┼──────┼──────────────────────┼───────┼────────┼─────────┼───────────────────────┼────────┼────┼────────┼────────────┼────────┼────────┼────────┤

```

#### AtomicBitset with bulk  set/reset
```text
┌──────────────────benchmark:──────────────────┬──────┬──CPU branch misses───┬─ops/s─┬──mean──┬─<CPU-I>─┬───CPU cache misses────┬─stddev─┬─#N─┬─CTX-SW─┬─total time─┬──min───┬─median─┬──max───┐
│                                              │ SKIP │                      │       │        │         │                       │        │    │        │            │        │        │        │
│    POSIX: 1 producers -< 1  >-> 1 consumers  │ PASS │ 68.8k / 591k = 11.6% │ 27.7M │  36 ms │    345m │  18.1k / 107k = 17.0% │ 496 us │ 8  │   104  │     289 ms │  35 ms │  36 ms │  37 ms │
│    POSIX: 1 producers -< 1  >-> 2 consumers  │ PASS │ 80.2k / 625k = 12.8% │ 18.0M │  55 ms │    366m │  62.2k / 165k = 37.8% │   6 ms │ 8  │   111  │     444 ms │  42 ms │  59 ms │  61 ms │
│    POSIX: 1 producers -< 1  >-> 4 consumers  │ PASS │ 80.6k / 624k = 12.9% │ 11.2M │  89 ms │    364m │  60.1k / 165k = 36.4% │   6 ms │ 8  │   120  │     712 ms │  79 ms │  90 ms │  95 ms │
│    POSIX: 2 producers -< 1  >-> 1 consumers  │ PASS │ 85.4k / 641k = 13.3% │  6.4M │ 157 ms │    373m │  82.3k / 198k = 41.5% │  21 ms │ 8  │   132  │       1  s │ 130 ms │ 145 ms │ 184 ms │
│    POSIX: 2 producers -< 1  >-> 2 consumers  │ PASS │ 92.0k / 691k = 13.3% │  6.1M │ 163 ms │    405m │  91.8k / 222k = 41.3% │  20 ms │ 8  │   137  │       1  s │ 143 ms │ 168 ms │ 191 ms │
│    POSIX: 2 producers -< 1  >-> 4 consumers  │ PASS │  113k / 851k = 13.3% │  6.1M │ 165 ms │    497m │   107k / 252k = 42.3% │  14 ms │ 8  │   161  │       1  s │ 137 ms │ 168 ms │ 183 ms │
│    POSIX: 4 producers -< 1  >-> 1 consumers  │ PASS │  110k / 805k = 13.6% │  5.6M │ 177 ms │    468m │   112k / 263k = 42.7% │  12 ms │ 8  │   169  │       1  s │ 163 ms │ 177 ms │ 198 ms │
│    POSIX: 4 producers -< 1  >-> 2 consumers  │ PASS │ 84.6k / 733k = 11.5% │  5.5M │ 183 ms │    427m │  12.1k / 137k =  8.8% │   4 ms │ 8  │   147  │       1  s │ 178 ms │ 183 ms │ 189 ms │
│    POSIX: 4 producers -< 1  >-> 4 consumers  │ PASS │ 124k / 1.09M = 11.3% │  3.6M │ 281 ms │    635m │  8.88k / 222k =  4.0% │   7 ms │ 8  │   248  │       2  s │ 266 ms │ 283 ms │ 291 ms │
├──────────────────────────────────────────────┼──────┼──────────────────────┼───────┼────────┼─────────┼───────────────────────┼────────┼────┼────────┼────────────┼────────┼────────┼────────┤
│    POSIX: 1 producers -<1024>-> 1 consumers  │ PASS │ 24.7k / 230k = 10.8% │ 23.5G │  43 us │    134m │ 1.38k / 4.44k = 31.2% │  43 us │ 8  │     1  │     341 us │  26 us │  27 us │ 156 us │
│    POSIX: 1 producers -<1024>-> 2 consumers  │ PASS │ 32.7k / 301k = 10.9% │ 10.1G │  99 us │    176m │ 1.23k / 6.12k = 20.0% │ 150 us │ 8  │     3  │     793 us │  39 us │  43 us │ 495 us │
│    POSIX: 1 producers -<1024>-> 4 consumers  │ PASS │ 51.8k / 448k = 11.6% │  3.9G │ 254 us │    262m │ 1.36k / 14.4k =  9.5% │ 211 us │ 8  │    12  │       2 ms │ 155 us │ 156 us │ 799 us │
│    POSIX: 2 producers -<1024>-> 1 consumers  │ PASS │ 58.7k / 508k = 11.5% │  1.2G │ 864 us │    297m │ 1.78k / 29.5k =  6.1% │ 150 us │ 8  │    33  │       7 ms │ 805 us │ 808 us │   1 ms │
│    POSIX: 2 producers -<1024>-> 2 consumers  │ PASS │ 56.8k / 493k = 11.5% │  1.2G │ 869 us │    288m │ 1.59k / 28.9k =  5.5% │ 148 us │ 8  │    33  │       7 ms │ 806 us │ 808 us │   1 ms │
│    POSIX: 2 producers -<1024>-> 4 consumers  │ PASS │ 55.7k / 484k = 11.5% │  1.1G │ 912 us │    283m │ 2.68k / 30.8k =  8.7% │ 180 us │ 8  │    34  │       7 ms │ 805 us │ 810 us │   1 ms │
│    POSIX: 4 producers -<1024>-> 1 consumers  │ PASS │ 55.5k / 483k = 11.5% │  1.2G │ 856 us │    282m │ 1.64k / 29.9k =  5.5% │ 154 us │ 8  │    33  │       7 ms │ 775 us │ 806 us │   1 ms │
│    POSIX: 4 producers -<1024>-> 2 consumers  │ PASS │ 63.8k / 554k = 11.5% │  700M │   1 ms │    323m │ 2.13k / 42.5k =  5.0% │ 422 us │ 8  │    45  │      11 ms │   1 ms │   1 ms │   3 ms │
│    POSIX: 4 producers -<1024>-> 4 consumers  │ PASS │ 70.9k / 615k = 11.5% │  729M │   1 ms │    359m │ 1.67k / 39.1k =  4.3% │ 238 us │ 8  │    43  │      11 ms │   1 ms │   1 ms │   2 ms │
├──────────────────────────────────────────────┼──────┼──────────────────────┼───────┼────────┼─────────┼───────────────────────┼────────┼────┼────────┼────────────┼────────┼────────┼────────┤

```